Move darwin specific test to separate file
