+++ pkg/networks/commands_test.go
@@ -21,24 +21,6 @@ func TestCheck(t *testing.T) {
 	assert.ErrorContains(t, err, "not defined")
 }
 
-func TestVDESock(t *testing.T) {
-	config, err := DefaultConfig()
-	assert.NilError(t, err)
-
-	vdeSock := config.VDESock("foo")
-	varRunDir := filepath.Join("/", "private", "var", "run", "lima")
-	assert.Equal(t, vdeSock, filepath.Join(varRunDir, "foo.ctl"))
-}
-
-func TestPIDFile(t *testing.T) {
-	config, err := DefaultConfig()
-	assert.NilError(t, err)
-
-	pidFile := config.PIDFile("name", "daemon")
-	varRunDir := filepath.Join("/", "private", "var", "run", "lima")
-	assert.Equal(t, pidFile, filepath.Join(varRunDir, "name_daemon.pid"))
-}
-
 func TestLogFile(t *testing.T) {
 	config, err := DefaultConfig()
 	assert.NilError(t, err)
